Add client-side JavaScript form validation
Validation messages and styling are triggered via the blur event, so as
not to annoy users.

A green tick mark is displayed inline if the input is considered valid.
Otherwise, a red symbol and help message and is shown below the field.
These changes are announced politely to screen readers.

Current validation messages are the browser defaults.